Move Noptester to fuzz utils
